Solving the CartPole-v0 environment through cases of the speed of the cartpole and the angle w.r.t. the vertical.
We will solve the environment by using the observations: if the cart is
moving to the left at a speed higher than a threshold, then move to the
right; if lower, move to the left. Likewise, if the angle of the pole
with the vertical is higher than a threshold, then move to the right; if
lower, move to the left.